Renamed editProfile to AdminPortalEditProfile.
