+++ front-end-203/src/App.js
@@ -6,12 +6,11 @@ import FlightSearch from "./flight-search/FlightSearch";
 import SignIn from "./signin-screen/SignInPage";
 import SignUp from "./signup-screen/SignUpPage";
 
-// TODO: This should be made common later on. But wait till after Jared does his updates and makes the nav bar component.
-import EditProfile from "./admin-portal/EditProfile";
 
 // Admin Portal
 import AdminPortalLogin from "./admin-portal/AdminPortalLogin";
 import AdminPortalHomePage from "./admin-portal/AdminPortalHomePage";
+import AdminPortalEditProfile from "./admin-portal/AdminPortalEditProfile";
 import AdminPortalManagePlane from "./admin-portal/AdminManagePlane";
 import AdminPortalEditPlane from "./admin-portal/AdminEditPlane";
 
@@ -24,12 +23,11 @@ function App() {
         <Route path="/signin" element={<SignIn />} />
         <Route path="/signup" element={<SignUp />} />
 
-        {/* TODO: This should be moved out of admin pages and then made generic when you click on the profile icon on the header bar */}
-        <Route path="/editprofile" element={<EditProfile />} />
 
         {/* ADMIN PORTAL */}
-        <Route path="/adminPortal/home" element={<AdminPortalHomePage />} />
         <Route path="/adminPortal/login" element={<AdminPortalLogin />} />
+        <Route path="/adminPortal/home" element={<AdminPortalHomePage />} />
+        <Route path="/adminPortal/editProfile" element={<AdminPortalEditProfile />} />
         <Route path="/adminPortal/planes" element={<AdminPortalManagePlane />} />
         <Route path="/adminPortal/planes/edit" element={<AdminPortalEditPlane />} />
       </Routes>
